punching bag scarecrows show you which room you havent been to and shake on player being damaged

+++ scripts/general/global.gd
@@ -369,14 +369,15 @@ func transition_to_world(destination_world_n: int, length = 1):
 	current_room_node.save_room()
 
 
-func return_to_menu():
+func return_to_menu(delete = false):
 	TransitionScreen.transition(3)
 	await TransitionScreen.on_transition_finished
-	call_deferred("restart_game_and_delete")
+	call_deferred("restart_game_and_delete", delete)
 
-func restart_game_and_delete():
+func restart_game_and_delete(delete):
 	get_node("/root/Game").queue_free()
-	SaverLoader.delete_save(game_save_id)
+	if delete:
+		SaverLoader.delete_save(game_save_id)
 	game_begun = false
 	game_save_id = 1
 	get_node("/root").add_child(load("res://scenes/general/main_menu.tscn").instantiate())
